Splitting RRWebEvent file
