+++ Sources/Swift/Integrations/SessionReplay/RRWeb/SentryRRWebMetaEvent.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@objcMembers
+class SentryRRWebMetaEvent: SentryRRWebEvent {
+    init(timestamp: Date, height: Int, width: Int) {
+        super.init(type: .meta, timestamp: timestamp, data: ["href": "", "height": height, "width": width])
+    }
+}
